com.unity.netcode@1.0.10 ## [1.0.10] - 2023-05-23
### Added

* What's New and Upgrade Guide section in the docs.
* New NetworkRequestListenResult cleanup component, that can be used to track the result of a listen request.

### Changed

* documentation index page with up-to-date info and links.
* Removed forcing local client/server to alway use the loopback address to connect.
* It is now possible to listen to the NetworkEndPoint.Any also for IPC connection.
* The NetworkStreamDriver.GetRemoteAddress always return a consistent address for the connection when the NetworkDriver is configured to use the Unity Relay. Before, an invalid address was returned after the connection has been established, that was incorrect.
* Exposed all the internal state of the NetworkTimeSystem as public API

### Fixed

* exceptions when NetworkRequestListen and/or. NetworkRequestConnect are handled and proper handling of multiple (erroneous) requests presents.
* A problem with InterpolatedTick, going back and not recovering correctly in presence of large application, either the server or the client, stalls (i.e after loading).

+++ Runtime/ClientServerWorld/ClientServerTickRate.cs
@@ -183,9 +183,9 @@ namespace Unity.NetCode
         public int MaxPredictionStepBatchSizeFirstTimeTick;
         /// <summary>
         /// Multiplier used to compensate received snapshot rate jitter when calculating the Interpolation Delay.
-        /// Default Value: 3.
+        /// Default Value: 1.25.
         /// </summary>
-        public uint InterpolationDelayJitterScale;
+        public float InterpolationDelayJitterScale;
         /// <summary>
         /// Used to limit the maximum InterpolationDelay changes in one frame, as percentage of the frame deltaTicks.
         /// Default value: 10% of the frame delta ticks. Smaller values will result in slow adaptation to the network state (loss and jitter)
